Optimized the map loading and printing
The map loading and printing is now fully seperate.
The Vector2 struct and EntityCharacters have also been changed to work
with the new map system.
The only change needed for the map system now is to make it work with
the 1D index instead of the 2D x,y vector.

+++ Header Files/Entities.h
@@ -20,10 +20,12 @@ namespace entities
 		Entity(std::string _nameIn, int _attackIn, int _mAttackIn, int _agilityIn, int _defenseIn, int _mDefenseIn, int _maxHealtIn, sprites::EntityCharacters _spriteDataIn, int _spawnPosXIn, int _spawnPosYIn);
 		//TODO: Update these to the new coordinate system.
 		sprites::EntityCharacters SpriteData;
-		game_utility::Vector2<int> SpawnPosition = { 0,0 }; 
-		game_utility::Vector2<int> PreviousPosition = { 0,0 };
-		game_utility::Vector2<int> CurrentPosition = { 0,0 };
-		game_utility::Vector2<int> CurrentFacingTile = { 0,0 };
+		game_utility::Vector2<int8_t> SpawnPosition = { 0,0 }; 
+		game_utility::Vector2<int8_t> PreviousPosition = { 0,0 };
+		game_utility::Vector2<int8_t> CurrentPosition = { 0,0 }; 
+		game_utility::Vector2<int8_t> CurrentFacingTile = { 0,0 };
+		//CurrentPosition, PreviousPosition and CurrentFacingTile are used with direction 
+		// so they need to be able to work with negative numbers.
 
 		virtual void Update() {};
 		virtual void interact() {};
